[Fix] #56 - PartCategory Core 모듈로 분리, 적용
