+++ SOPT-iOS/Projects/Modules/Network/Sources/Entity/PushAlarmSettingEntity.swift
@@ -8,6 +8,8 @@
 
 import Foundation
 
+import Core
+
 // MARK: - PushAlarmSettingEntity
 public struct PushAlarmSettingEntity: Codable {
     public let alerts: [Alert]
@@ -19,12 +21,12 @@ public struct Alert: Codable {
     public let isAlert: Bool
 }
 
-public enum PartCategory: String, Codable {
-    case fullNotice = "all"
-    case plan = "plan"
-    case design = "design"
-    case ios = "iOS"
-    case android = "android"
-    case web = "web"
-    case server = "server"
-}
+//public enum PartCategory: String, Codable {
+//    case fullNotice = "all"
+//    case plan = "plan"
+//    case design = "design"
+//    case ios = "iOS"
+//    case android = "android"
+//    case web = "web"
+//    case server = "server"
+//}
